{"schema":"decodex/commit/1","summary":"Modularize live frame store","authority":"manual"}
